wordle attempt

+++ wordle.cpp
@@ -13,7 +13,7 @@ using namespace std;
 
 
 // Add prototypes of helper functions here
-void wordleHelper(const std::string& in, const std::string& floating, const std::set<std::string>& dict, int idxIn, std::map<char, int>& floatingUsage, std::set<std::string>& answerBank, std::string& tempString, int floatingRemaining, int blanks);
+void wordleHelper(const std::string& in, const std::string& floating, const std::set<std::string>& dict, size_t idxIn, std::map<char, int>& floatingUsage, std::set<std::string>& answerBank, std::string& tempString, int floatingRemaining, int blanks);
 
 // Definition of primary wordle function
 std::set<std::string> wordle(
